+++ Modules/antibodies/Controller/OptionController.php
@@ -55,13 +55,13 @@ class OptionController extends AntibodiesController {
         if($form->check()){
             $name = $this->request->getParameter("nom");
             if (!$id){
-                $this->model->add($name, $id_space);
+                $id = $this->model->add($name, $id_space);
             }
             else{
                 $this->model->edit($id, $name, $id_space);
             }
             
-            $this->redirect("option/".$id_space);
+            return $this->redirect("option/".$id_space, [], ['option' => ['id' => $id]]);
         }
 
         $this->render(array(