+++ Modules/antibodies/Model/Linker.php
@@ -83,6 +83,7 @@ class Linker extends Model {
         $sql = "insert into ac_linkers(nom, id_space)"
                 . " values(?,?)";
         $this->runRequest($sql, array($name, $id_space));
+        return $this->getDatabase()->lastInsertId();
     }
 
     /**