+++ Modules/antibodies/Model/Source.php
@@ -93,6 +93,7 @@ class Source extends Model {
         $sql = "insert into ac_sources(nom, id_space)"
                 . " values(?,?)";
         $this->runRequest($sql, array($name, $id_space));
+        return $this->getDatabase()->lastInsertId();
     }
 
     public function importSource($id, $name, $id_space) {